feat: made --output work with --dir and --batch and cleanup
